Removed renderer component and moved rendering code to ObserverFrame
