--- AGXUnity/Rendering/ObserverFrameRenderer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 05672516ee16bea43b20146d8c4c11e5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
