--- AGXUnity/Rendering/ObserverFrameRendererManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bc92fda39e0566243a8e70ac36ed4f9f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
