--- Editor/CustomEditors/AGXUnity+Rendering+ObserverFrameRendererEditor.cs
@@ -1,13 +1,0 @@
-
-using System;
-using AGXUnity;
-using AGXUnity.Collide;
-using UnityEditor;
-
-namespace AGXUnityEditor.Editors
-{
-  [CustomEditor( typeof( AGXUnity.Rendering.ObserverFrameRenderer ) )]
-  [CanEditMultipleObjects]
-  public class AGXUnityRenderingObserverFrameRendererEditor : InspectorEditor
-  { }
-}
