--- Editor/CustomEditors/AGXUnity+Rendering+ObserverFrameRendererEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1fb6b03e3aedad640bfb6440ad07e6a5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
