--- Editor/CustomEditors/AGXUnity+Rendering+ObserverFrameRendererManagerEditor.cs
@@ -1,13 +1,0 @@
-
-using System;
-using AGXUnity;
-using AGXUnity.Collide;
-using UnityEditor;
-
-namespace AGXUnityEditor.Editors
-{
-  [CustomEditor( typeof( AGXUnity.Rendering.ObserverFrameRendererManager ) )]
-  [CanEditMultipleObjects]
-  public class AGXUnityRenderingObserverFrameRendererManagerEditor : InspectorEditor
-  { }
-}
