--- Editor/CustomEditors/AGXUnity+Rendering+ObserverFrameRendererManagerEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f72a69f446077ff498e6c1dcfb4ccf19
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
